--- a/tests/test_utils/test_build.lua
+++ b/tests/test_utils/test_build.lua
@@ -2,7 +2,7 @@ local test_build = {}
 
 function test_build:build(argv)
     os.exec("xmake f -c -D -y")
-    os.exec("xmake")
+    os.exec("xmake -D")
 end
 
 function main()