… xmake/modules/private/action/build/target.lua
… xmake/modules/private/service/distcc_build/client.lua