- [x] Reduce the time spent on coroutine scheduling
- [x] Add `build.distcc.remote_only` policy to enable build on only remote machines

### Before optimization

```console
ruki:xmake ruki$ xmake l tests/benchmarks/async/runjobs.lua
runjobs(10000/1): 1678 ms, plain: 0 ms
runjobs(10000/10): 1853 ms, plain: 0 ms
runjobs(10000/100): 1904 ms, plain: 1 ms
runjobs_proc(1000/10): 63061 ms
```

### After optimization

```console
ruki:xmake ruki$ xmake l tests/benchmarks/async/runjobs.lua
runjobs(10000/1): 50 ms, plain: 0 ms
runjobs(10000/10): 51 ms, plain: 0 ms
runjobs(10000/100): 67 ms, plain: 0 ms
runjobs_proc(1000/10): 52977 ms
```